Enforce clang-tidy readability-implicit-bool-conversion across codebase
Applies clang-tidy-18 readability-implicit-bool-conversion autofixes across
the C++ codebase and adds the check to .clang-tidy-enforce. This aligns with
the house rule of comparing integer/floating-point/pointer values explicitly
rather than relying on implicit boolean casts.

Generated-by: Claude Opus 4.7 (1M context) <noreply@anthropic.com>

+++ cpp/core/math/fft.h
@@ -71,7 +71,7 @@ inline std::string direction_str(int direction) { return (direction == FFTW_FORW
 
 inline size_t shift(size_t pos, const size_t size, const bool centre_FFT, const bool inverse) {
   if (centre_FFT) {
-    pos += (size + 1 + (!inverse)) / 2;
+    pos += (size + 1 + static_cast<size_t>(!inverse)) / 2;
     if (pos >= size)
       pos -= size;
   }
